requirements.txt
argparse
pyyaml
jsbeautifier
beautifulsoup4
requests
tqdm
pycryptodomex
undetected-chromedriver
pyrogram
tgcrypto
python-dotenv